refactor: Coupon 필드 데이터타입 변경, @Repository 어노테이션 추가
- Counpon 엔티티의 discountRate과 quantity는 int데이터타입 범위로 충분하기 때문에 Long에서 int로 변경
- Repository에 @Repository 어노테이션을 추가하여 빈으로 등록

+++ src/main/java/com/example/coupon/domain/Coupon.java
@@ -17,19 +17,19 @@ public class Coupon {
     private Long id;
 
     @Column(nullable = false)
-    private Long quantity;
+    private int quantity;
 
     @Column(nullable = false, length = 255)
     private String name;
 
     @Column(nullable = false)
-    private Long discountRate;
+    private int discountRate;
 
     @Column(nullable = false)
     private LocalDate expirationDate;
 
     @Builder
-    public Coupon(Long quantity, String name, Long discountRate, LocalDate expirationDate) {
+    public Coupon(int quantity, String name, int discountRate, LocalDate expirationDate) {
         this.quantity = quantity;
         this.name = name;
         this.discountRate = discountRate;
